+++ app/models/project.rb
@@ -4,6 +4,7 @@ class Project < ApplicationRecord
   attr_writer :establishment, :incoming_trust
 
   delegated_type :tasks_data, types: %w[Conversion::TasksData, Transfer::TasksData], dependent: :destroy
+  delegate :local_authority_code, to: :establishment
 
   has_many :notes, dependent: :destroy
   has_many :contacts, dependent: :destroy, class_name: "Contact::Project"