+++ config/routes.rb
@@ -87,6 +87,9 @@ Rails.application.routes.draw do
             get "/", to: "projects#index"
             get ":trust_ukprn", to: "projects#show", as: :by_trust
           end
+          namespace :local_authorities, path: "local-authorities" do
+            get "/", to: "projects#index"
+          end
         end
         namespace :team, path: "team" do
           get "in-progress", to: "projects#in_progress"